Reorganization
This looks like a lot of changes, but it's really just moving pretty much all the game files into different folders to match a more consistent structure to the new architecture.
