+++ game/objs/espionage.py
